docs(cascader): update demo code
re #413
